+++ Makefile
@@ -72,6 +72,7 @@ INSTALL_PYLIB_FILES = \
     _stbt/lmdb/cpython.so \
     _stbt/lmdb/LICENSE \
     _stbt/logging.py \
+    _stbt/match.py \
     _stbt/motion.py \
     _stbt/ocr.py \
     _stbt/power.py \
