+++ _stbt/pylint_plugin.py
@@ -73,7 +73,7 @@ class StbtChecker(BaseChecker):
             if isinstance(node.parent, Expr):
                 for inferred in _infer(node.func):
                     if inferred.root().name in ('stbt', '_stbt.core',
-                                                '_stbt.ocr'):
+                                                '_stbt.match', '_stbt.ocr'):
                         self.add_message(
                             'E7002', node=node, args=node.func.as_string())
